Implemented extract
