+++ generator/src/commands/extract.rs
@@ -1,6 +1,9 @@
-use anyhow::Result;
+use anyhow::{Context, Result};
 use clap::Parser;
-use std::path::PathBuf;
+use std::{env, path::PathBuf};
+use tracing::*;
+
+use crate::ChipDescription;
 
 /// Extract all metadata information from the SVD and apply any transforms.
 #[derive(Parser)]
@@ -9,13 +12,46 @@ pub struct Extract {
     #[clap(required = true)]
     pub chip: String,
     /// Output directory of the metadata.
-    #[clap(short, long, default_value = "./data/metadata/raw")]
-    pub output: PathBuf,
+    #[clap(short, long)]
+    pub output: Option<PathBuf>,
 }
 
 /// Command to extract all metadata information from a single SVD.
 ///
 /// Applies any transforms automagically.
 pub fn extract(extract: Extract) -> Result<()> {
-    todo!()
+    let chip = crate::CHIPS
+        .iter()
+        .find(|chip_description| chip_description.chip == extract.chip)
+        .context("selected chip does not exist in generate list")?;
+
+    extract_chip(chip, extract.output)
+}
+
+fn extract_chip(chip_description: &ChipDescription, output: Option<PathBuf>) -> Result<()> {
+    let current_dir = env::current_dir()?;
+
+    let chip_src_dir = current_dir
+        .join("data")
+        .join("mcux-soc-svd")
+        .join(chip_description.chip);
+
+    for core in chip_description.cores {
+        let svd = chip_src_dir.join(core).with_extension("xml");
+        debug!("svd path: {:?}", svd);
+        let transforms_dir = current_dir.join("data").join("transforms");
+        debug!("transforms path: {:?}", transforms_dir);
+        let output_dir = output.clone().unwrap_or_else(|| {
+            current_dir
+                .join("data")
+                .join("metadata")
+                .join("peripherals")
+                .join("raw")
+                .join(core)
+        });
+
+        crate::metadata::extract_peripherals(&svd, core, &transforms_dir, &output_dir)?;
+    }
+
+    Ok(())
 }
